test: fix deadlock in unit test
The current unit test still occasionally deadlocks. Which is caused by
a global mutex in `std::sys_common::backtrace::print()`.
The root cause of deadlocks is the use of `fork()` in a multi-threaded
test environment. In this commit we limit the cargo test to run one unit
test at a time by setting `--test-threads=1`.